backend/pyproject.toml
[project]
name = "caria-backend"
version = "0.1.0"
description = "CARIA AI Elderly Care Companion — ADK Agent Backend"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.26",
    "firebase-admin>=6.6",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "playwright>=1.49",
    "google-cloud-storage>=2.19",
    "google-cloud-aiplatform>=1.78",
    "google-cloud-firestore>=2.20",
    "google-genai>=1.0",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
